update llama llama2
